+++ src/main/java/application/modele/Ennemie.java
@@ -23,12 +23,12 @@ public class Ennemie extends Personnage {
     }
 
     @Override
-    public float getX() {
+    public double getX() {
         return x;
     }
 
     @Override
-    public float getY() {
+    public double getY() {
         return y;
     }
